Update: move charge and connectivity to a new module
