+++ enzy_htp/quantum/api.py
@@ -21,7 +21,7 @@ from enzy_htp.structure import (
     StructureConstraint
 )
 
-from enzy_htp.structure.structure_operation import init_charge
+from enzy_htp.structure.structure_enchantment import init_charge
 from enzy_htp.chemical import QMLevelOfTheory, LevelOfTheory
 from enzy_htp.core.logger import _LOGGER
 import enzy_htp.core.job_manager as armer
