+++ enzy_htp/structure/structure_operation/__init__.py
@@ -1,4 +1,6 @@
-"""the submodule for operation of Structure objects
+"""the submodule for operation of Structure objects. This module should just rely on Structure and below
+so that it can be used by anywhere that uses Structure.
+It is seperated from the Structure class just for maintenance need. (i.e.: seperate the DS code and the operation code)
 
 Author: Qianzhen (QZ) Shao, <shaoqz@icloud.com>
 Date: 2022-09-19
@@ -18,9 +20,3 @@ from .residue import (
     remove_side_chain_mutating_atom,
     check_res_topology_error,
 )
-
-from .connectivity import (
-    init_connectivity
-)
-
-from .charge import init_charge
