+++ test/_interface/test_multiwfn_interface.py
@@ -15,7 +15,7 @@ from enzy_htp._config.armer_config import ARMerConfig
 from enzy_htp.electronic_structure import ElectronicStructure
 from enzy_htp.structure import structure_constraint as stru_cons
 from enzy_htp.structure.structure_region import create_region_from_selection_pattern
-from enzy_htp.structure.structure_operation.charge import init_charge
+from enzy_htp.structure.structure_enchantment.charge import init_charge
 from enzy_htp import PDBParser
 from enzy_htp import interface
 from enzy_htp._interface.gaussian_interface import GaussianSinglePointEngine
